3% speed gain by only calling glColorPointer when needed
This commit may introduce potentially some color regressions with the 3 vertices of the triangle, if changing the circle precision from 180 to something lower (especially if it isn't a multiple of 12), but we can always force a constant synchronization by calling synchronise_colors() under the condition `if (true)` instead of detecting an interval.

Until then, the speed gain from not constantly uploading the new colors[] array with glColorPointer on every single frame, seems worth it.