canceled my implementation of glRectf()
At this point in time, the function only conflicts with the required functions this wrapper does actually need to have.  This function was perfectly working back when I first implemented it in this wrapper, but the wrapper has since come a long way and has kept more accurate track of the virtual GL state machine than I would have hoped to accomplish.  With the unexpected correctness that I have achieved with this wrapper, maintaining a correct glRectf() emulation would require either:
a) setting up a detachment and restoriation of the old GLSL vertex program while temporarily using a new one to avoid conflicting with the primary dummy shader program emulation we do
b) resorting to glClear(GL_COLOR_BUFFER_BIT) with glScissor() and scissor test enabled, which requires at least as much state machine probing

It is unfortunate that glRect* was deprecated along with GL_QUADS, since it can very well have been faster than forcing triangles on every single video card (or software implementations).

The other reason why I am removing glRect* instead of caring to fight over my new wrapper implementation over how to keep it working at this point in time is that it assumes only "flat" rectangles.

A rectangle can be slanted or diagonal and still be a rectangle with 4 90-degree angles, and that's exactly the kind of "slanted rectangle" that glRect*() does not allow us to specify.  Therefore, it's not geometrically correct in every sense, so I don't really care that much for it at this point.