Manually force every vertex attribute ID binding.
By default, GLSL shader linkage has the compiled vertex attribute IDs assigned randomly and automatically to any number (for example, *sometimes* 0 for the vertex position attribute's ID).

This makes it extremely difficult for me to guarantee that I am binding and buffering new vertex array data to the correct ID every single time, in a well-defined manner for every browser on every platform.

This change should immensely help the stability of the program and prevent more bugs in advance later on down the road as I continue with more fundamentally structural changes to the entire wrapper.