Major rewrites to switch pre-linked programs, not re-compile.
This has both a static performance advantage (as observed in the JavaScript profiler that comes with Slackware Linux FireFox on my NVIDIA 8 series) and a memory space safety advantage.

As observed on said distro, the memory leak would accumulate more and more cruft each time a shader was re-compiled for re-linking the vertex program.  I was able to remove this overhead after a week of frustration with overhauling the code structure and fixing regressions.

Additionally, the global variable `dummy_scripts[]` is removed, and all JSLint string offset warnings are fixed.