color morphing/rotation animation + big speed gain
Now that both glVertexPointer and glColorPointer are moved out of the main `display()` render loop and into the initialization function, we no longer have to worry about the bottleneck of gl[Sub]BufferData.

This does lose us the ability to force constant triangle vertex colors by not constantly refreshing them to exactly red, exactly green and exactly blue, but I think I like the new color rotation effect for the moment and have intentionally featured it in to take advantage of that semi-desired side effect.