Rotate the triangle clock-wise, not counter-clock-wise.
This just came to me while I was watching the new color morphing effect of the 3 vertex colors--as they cycle along the colors of the rainbow, rotating counter-clock-wise (as if measured by a protractor in degrees) goes backwards from red, to purple, to blue, to cyan, green, to yellow and back to red.  Reversing the order of the index buffer elements is enough to satisfy this condition as a clock-wise rotation.