renamed GL_initialise to GL_get_context
One of the reasons why I didn't like the "initialize" name was that the nature of this wrapper kept track of the WebGL context using a global variable, which could collide if multiple WebGL contexts are open using this same page to emulate C-style OpenGL functions.

So it's not really "initializing"; it's just fetching the current status of the wrapper's stored emulated context.