Replace spaces with tabs for GLSL script indentations.
It's not as if tabs are any more correct than using spaces to indent, but I am going to be looking at maintaining a minified or compressed version of gl.js that programmers can reference for faster download times.  Minification of the ES6 `multi-
line
string
feature`
cannot safely remove whitespace from these parts of the minified JS, and the GLSL scripts were all using four-space indents instead of tabs just to make the minified result look even more incomplete.  I am not into using tabs over spaces (especially with C-based languages like GLSL), but in this case I feel it is wiser to do so than to not do so.